requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
cryptography==2.7
defusedxml==0.6.0
lxml==4.4.1
paramiko==2.6.0
pkg-resources==0.0.0
pycparser==2.19
PyNaCl==1.3.0
python-gvm==1.0.0
six==1.12.0
